README.md
This project aims us to predict the count of bike rented using historical data over environmental and seasonal settings, so that the entities can oversee these systems to manage the required bikes and arrange them in cost-effective way based on environmental and seasonal factors. The data we are provided with consists of 15 variables where 15th variable ‘cnt’ is the count of Bike users. Our task is to build regression models to predict the daily count of bike rented based on weather and season. We split the data into train and test, then build the model using various regression models such as linear regression, decision tree and Random forest. These models are evaluated based on regression metrics such as rmse, mape, mae and mse. The rmse and mape value is less in random forest hence the accuracy of the model is more compared to other models. Random Forest model is deployed and predict the test case (count of bike rental).